requirements.txt
selenium==3.141.0
requests==2.22.0
Pillow==6.1.0
beautifulsoup4==4.8.0
